Change project package name from rick_and_morty_list to ramlist
