+++ app/src/main/java/com/pnow/ramlist/app/RickAndMortyApp.kt
@@ -1,0 +1,7 @@
+package com.pnow.ramlist.app
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class RickAndMortyApp : Application()
