+++ app/src/main/java/com/pnow/ramlist/app/data/di/module/AppModule.kt
@@ -1,0 +1,27 @@
+package com.pnow.ramlist.app.data.di.module
+
+import android.content.Context
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.qualifiers.ApplicationContext
+import dagger.hilt.components.SingletonComponent
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.Dispatchers
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object AppModule {
+
+    @Provides
+    @Singleton
+    fun provideContext(@ApplicationContext appContext: Context): Context {
+        return appContext
+    }
+
+    @Provides
+    @Singleton
+    fun provideDispatcher(): CoroutineDispatcher = Dispatchers.Main
+
+}
