+++ app/src/main/java/com/pnow/ramlist/app/data/model/LocationModel.kt
@@ -1,0 +1,14 @@
+package com.pnow.ramlist.app.data.model
+
+import com.google.gson.annotations.SerializedName
+
+data class LocationModel(
+    @field:SerializedName("id")
+    val id: Int,
+    @field:SerializedName("name")
+    val name: String,
+    @field:SerializedName("type")
+    val type: String,
+    @field:SerializedName("dimension")
+    val dimension: String
+)
