+++ app/src/main/java/com/pnow/ramlist/app/data/networking/model/CharacterResponseModel.kt
@@ -1,0 +1,5 @@
+package com.pnow.ramlist.app.data.networking.model
+
+import com.pnow.ramlist.app.data.model.character.CharacterModel
+
+data class CharacterResponseModel(val results: List<CharacterModel>)
