+++ app/src/main/java/com/pnow/ramlist/app/ui/model/EpisodeUIModel.kt
@@ -1,0 +1,9 @@
+package com.pnow.ramlist.app.ui.model
+
+
+data class EpisodeUIModel(
+    val id: Int,
+    val name: String,
+    val episodeNumber: String,
+    val date: String,
+)
