+++ app/src/main/java/com/pnow/ramlist/app/ui/model/LocationUIModel.kt
@@ -1,0 +1,8 @@
+package com.pnow.ramlist.app.ui.model
+
+
+data class LocationUIModel(
+    val name: String,
+    val type: String,
+    val dimension: String
+)
