--- app/src/main/java/com/pnow/rick_and_morty_list/app/RickAndMortyApp.kt
@@ -1,7 +1,0 @@
-package com.pnow.rick_and_morty_list.app
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class RickAndMortyApp : Application()
