--- app/src/main/java/com/pnow/rick_and_morty_list/app/data/di/module/AppModule.kt
@@ -1,27 +1,0 @@
-package com.pnow.rick_and_morty_list.app.data.di.module
-
-import android.content.Context
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.qualifiers.ApplicationContext
-import dagger.hilt.components.SingletonComponent
-import kotlinx.coroutines.CoroutineDispatcher
-import kotlinx.coroutines.Dispatchers
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object AppModule {
-
-    @Provides
-    @Singleton
-    fun provideContext(@ApplicationContext appContext: Context): Context {
-        return appContext
-    }
-
-    @Provides
-    @Singleton
-    fun provideDispatcher(): CoroutineDispatcher = Dispatchers.Main
-
-}
