--- app/src/main/java/com/pnow/rick_and_morty_list/app/data/model/EpisodeModel.kt
@@ -1,14 +1,0 @@
-package com.pnow.rick_and_morty_list.app.data.model
-
-import com.google.gson.annotations.SerializedName
-
-data class EpisodeModel(
-    @field:SerializedName("id")
-    val id: Int,
-    @field:SerializedName("name")
-    val name: String,
-    @field:SerializedName("air_date")
-    val airDate: String,
-    @field:SerializedName("episode")
-    val episode: String
-)
