--- app/src/main/java/com/pnow/rick_and_morty_list/app/data/model/LocationModel.kt
@@ -1,14 +1,0 @@
-package com.pnow.rick_and_morty_list.app.data.model
-
-import com.google.gson.annotations.SerializedName
-
-data class LocationModel(
-    @field:SerializedName("id")
-    val id: Int,
-    @field:SerializedName("name")
-    val name: String,
-    @field:SerializedName("type")
-    val type: String,
-    @field:SerializedName("dimension")
-    val dimension: String
-)
