--- app/src/main/java/com/pnow/rick_and_morty_list/app/data/networking/model/CharacterResponseModel.kt
@@ -1,5 +1,0 @@
-package com.pnow.rick_and_morty_list.app.data.networking.model
-
-import com.pnow.rick_and_morty_list.app.data.model.character.CharacterModel
-
-data class CharacterResponseModel(val results: List<CharacterModel>)
