--- app/src/main/java/com/pnow/rick_and_morty_list/app/ui/model/EpisodeUIModel.kt
@@ -1,9 +1,0 @@
-package com.pnow.rick_and_morty_list.app.ui.model
-
-
-data class EpisodeUIModel(
-    val id: Int,
-    val name: String,
-    val episodeNumber: String,
-    val date: String,
-)
