--- app/src/main/java/com/pnow/rick_and_morty_list/app/ui/model/LocationUIModel.kt
@@ -1,8 +1,0 @@
-package com.pnow.rick_and_morty_list.app.ui.model
-
-
-data class LocationUIModel(
-    val name: String,
-    val type: String,
-    val dimension: String
-)
