+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,13 +7,13 @@
 
     <fragment
         android:id="@+id/characterList"
-        android:name="com.pnow.rick_and_morty_list.app.ui.fragment.CharactersListFragment"
+        android:name="ramlist.app.ui.fragment.CharactersListFragment"
         android:label="CharacterList"
         tools:layout="@layout/fragment_characters_list"/>
 
     <fragment
         android:id="@+id/characterDetails"
-        android:name="com.pnow.rick_and_morty_list.app.ui.fragment.CharacterDetailsFragment"
+        android:name="ramlist.app.ui.fragment.CharacterDetailsFragment"
         android:label="CharacterDetails"
         tools:layout="@layout/fragment_character_deatails"/>
